pyproject.toml
[project]
name = "Karboni"
description = "Mirror a Zotero library into a SQL database"
authors = [
    {name = "David Lesieur", email="karboni@whiskyechobravo.com"},
]
readme = "README.md"
keywords = ["academia", "zotero"]
license = "GPL-3.0-only"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Database :: Front-Ends",
    "Topic :: Education",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dynamic = ["version"]
requires-python = ">= 3.11"
dependencies = [
    "anyio >= 4.9.0",
    "click >= 8.2.1",
    "click-option-group >= 0.5.7",
    "colorlog >= 6.10.1",
    "python-dotenv >= 1.1.1",
    "httpx >= 0.28.1",
    "SQLAlchemy >= 2.0.41",
]

[project.optional-dependencies]
tests = [
    "coverage >= 6.1.2, < 8",
]
dev = [
    "httpx",
    "mypy >= 1.16.1",
    "pre-commit >= 4.2.0",
    "rope >= 1.13.0",
    "ruff >= 0.12.11",
]
docs = [
    "mkdocs-material >= 9.4, < 10",
    "mike >= 1.1.2, < 3",
]

[project.urls]
Documentation = "https://github.com/whiskyechobravo/karboni/blob/main/README.md"
Code = "https://github.com/whiskyechobravo/karboni"
Changes = "https://github.com/whiskyechobravo/karboni/blob/main/CHANGELOG.md"
Issues = "https://github.com/whiskyechobravo/karboni/issues"

[build-system]
requires = ["hatchling>=1.26.1", "hatch-vcs"]
build-backend = "hatchling.build"

[project.scripts]
karboni = "karboni.cli:main"

[tool.hatch.version]
source = "vcs"

[tool.coverage.run]
branch = true
source = ["karboni"]

[tool.coverage.paths]
source = [
    "src",
]

[tool.mypy]
python_version = "3.14"
strict = true
mypy_path = "src"
files = ["src"]

[[tool.mypy.overrides]]
module = ["anyio.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["httpx.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["sqlalchemy.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["click.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["click-option-group.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["dotenv.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
disallow_untyped_calls = false

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "F",    # Pyflakes
    "W",    # pycodestyle
    "E",    # pycodestyle
    "I",    # isort
    "N",    # pep8-naming
    "D",    # pydocstyle
    "UP",   # pyupgrade
    "YTT",  # flake8-2020
    "BLE",  # flake8-blind-except
    "A",    # flake8-builtins
    "COM",  # flake8-commas
    "T10",  # flake8-debugger
    "EM",   # flake8-errmsg
    "EXE",  # flake8-executable
    "INT",  # flake8-gettext
    "ISC",  # flake8-implicit-str-concat
    "ICN",  # flake8-import-conventions
    "INP",  # flake8-no-pep420
    "PIE",  # flake8-pie
    "T20",  # flake8-print
    "PYI",  # flake8-pyi
    "Q",    # flake8-quotes
    "RSE",  # flake8-raise
    "SLF",  # flake8-self
    "SLOT", # flake8-slots
    "TID",  # flake8-tidy-imports
    "TC",   # flake8-type-checking
    "ARG",  # flake8-unused-arguments
    "PTH",  # flake8-use-pathlib
    "PGH",  # pygrep-hooks
    "PL",   # Pylint
    "FLY",  # flynt
    "TRY",  # tryceratops
    "PERF", # Perflint
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "G",    # flake8-logging-format
    "RET",  # flake8-return
    "SIM",  # flake8-simplify
    "ERA",  # eradicate
    "RUF",  # Ruff-specific rules
]
ignore = [
    "D100",    # Missing docstring in public module.
    "D101",    # Missing docstring in public class.
    "D102",    # Missing docstring in public method.
    "D103",    # Missing docstring in public function.
    "D105",    # Missing docstring in magic method.
    "D106",    # Missing docstring in public nested class.
    "D107",    # Missing docstring in __init__.
    "D212",    # Multi-line docstring summary should start at the first line.
    "COM812",  # Missing trailing comma.
    "PLR0912", # Too many branches.
    "PLR0913", # Too many arguments in function definition.
    "PLR0915", # Too many statements.
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["karboni"]
